Fix XML file icons taking up 2 characters
Some symbols in a Nerd Font take up two characters in width when
displayed in a terminal; the XML file icon is one of them. This causes
the file names to be misaligned, since other icons take up one character
in width. This changes the icon to be "seti-html", which takes up one
character like usual.